[FIX] mail: messaging menu last message when receiving new message
Before this commit, when we received a new message, the counter was correctly
updated, but not the last message in the messaging menu preview.

This is due to making changes to the counter before adding the message in the
thread. When there is a change in the unread counter of a thread, it warns its
views. The messaging menu renders the previews of each threads in the preview
items, using in particular the last message in the thread.

This commit fixes the issue by making changes on the unread counter of a thread
after the message has been added to the thread.