[IMP] mail: Chatter: add 'display_log_button' option
This rev. allows to instantiate a Chatter Widget with option
'display_log_button' (which shows the 'Log note' button).

This option already existed, but it had to be specified in the
nodeOptions (i.e. in the view arch), which is not convenient when
a Chatter is instantiated outside the context of a FormView.

Part of task 1853490