[IMP] mail: Allow to send a notification to the author
Purpose
=======

Sometimes we could want to send a notification to the author too.

Like in the onboarding bars, where we send a sample invoice to the administrator.

He's the author of the message, but the goal is to send him an email to check the
full process when sending a sales order/invoice by email to a customer.

Specification
=============

For that purpose, if a key 'mail_notify_author' is set on the context when
sending the email, then notify the mail author.